Rename bot.py to twitchbot.py
